% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{convert_bool}
\alias{convert_bool}
\title{Convert R bool value into Python bool value as a string}
\usage{
convert_bool(bool)
}
\arguments{
\item{bool}{R bool value}
}
\value{
Python bool value as a string
}
\description{
\code{convert_bool} converts a R bool value (TRUE or FALSE) into a Python bool value (True or False). This function is not exported.
}
