% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{create_widget_id}
\alias{create_widget_id}
\title{Create widget id}
\usage{
create_widget_id(bytes = 10)
}
\arguments{
\item{bytes}{number to sample}
}
\value{
An id string
}
\description{
Create widget id if not exists.
}
