% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{resolve_sizing}
\alias{resolve_sizing}
\title{Resolve widget sizing policy}
\usage{
resolve_sizing(x, sp)
}
\arguments{
\item{x}{A widget object whose size is to be determined.}

\item{sp}{The sizing policy to use.}
}
\value{
A list that is guaranteed to have `width` and `height` values (number of CSS unit string).
}
\description{
Figure out what width/height to use, taking a widget object and sizing policy.
}
